Fixes to make NPM work:
Handle HTTP responses with no content better -- only 20x responses that
are not 204 should be assumed that they will contain content.

Handle more filesystem errors with the right codes rather than "EIO".
This needs to continue and needs to include Java 6.

Wrap tick execution in a JS function to prevent errors with anonymous
callbacks and to simplify the domain processing logic.