Use ephemeral ports for tests.
This addresses a cause of test errors in which our use of hard-coded
ports sometimes causes failures because those ports are in use.
Instead, always listen on port 0, then check back and see what
port was allocated before sending requests.

Making this work requires some changes in the HTTP adapter so that it
can report the correct address.

For adapters that can't (a servlet adapter, or Apigee Edge) we will
instead return "localhost" and port -1. That way clients that
call 'address' will at least get something they can parse but
can't actually connect to.