make HTTP adaptor handling exit/uncaught exceptions configurable
This commit uses two environment variables to make it where the HTTP
adaptor can handle `exit` and `uncaughtException` events like it did
prior to `0.9.0`.  Since `0.9.0` fixed a bug, this default behavior
of Trireme stays the same but this new feature does allow embedders
the ability to have backward compatibility prior to `0.9.0` when it
comes to this.

Fixes #78