Fix when Trireme Buffer has an offset set for its underlying array
`ByteBuffer.wrap()` was not previously setting the offset correctly
`Buffer` was setting the position for the `ByteBuffer` assuming it
meant offset.

Because [ByteBuffer]( https://docs.oracle.com/javase/7/docs/api/java/nio/ByteBuffer.html#wrap(byte[],%20int,%20int))
does not support setting the offset _"Its backing array will be
the given array, and its array offset will be zero."_. Because of
this we create an array copy of the usable array in `Buffer` and pass
that to `ByteBuffer.wrap`. This only applies to caes where an offset
is set to avoid unnecessary copies of the array.

This causes an error #181 for zlib compression when the Buffer comes
from the HttpParser since the body is just an offset of the underlying
full http message returned.

Cleans up failing dns test.